job_scraper-main/requirements.txt
beautifulsoup4==4.9.3
requests==2.26.0
pymongo==3.12.0
django==4.2.0
python-dotenv==0.19.0
pandas==1.3.3
djongo==1.3.6
pandas==2.2.0
numpy==1.26.4
pymongo==4.6.1
python-dotenv==1.0.1
matplotlib==3.8.2
seaborn==0.13.2
statistics==1.0.3.5